requirements.txt
click>=8.1.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=12.5.1
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0 